Accept suggestions from CodeRabbitAI
